Move NotFound from components to views
